requirements.txt
asttokens==3.0.0
blinker==1.9.0
boto3==1.35.87
botocore==1.35.87
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
executing==2.1.0
findspark==2.0.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Mail==0.10.0
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
py4j==0.10.9.7
Pygments==2.18.0
PyJWT==2.10.1
PyMySQL==1.1.1
pyspark==3.5.3
python-dateutil==2.9.0.post0
pywin32==308
pyzmq==26.2.0
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
spark-submit==1.4.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
